pyproject.toml
[project]
name = "django-kit-starter"
version = "0.1.0"
description = "Django Kit Starter"
authors = [
    {name = "Enrique Lazo Bello",email = "enlabe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-simple-history (>=3.10.1,<4.0.0)",
    "django-constance (>=4.3.2,<5.0.0)",
    "django-unfold (>=0.66.0,<0.67.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "django-money (>=3.5.4,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^6.0.0"
